add translate and shared module
