+++ src/app/modules/errors/error-404/error-404.template.html
@@ -12,7 +12,6 @@
 -->
 
 <div>
-   <!--<h1>{{ 'ERRORS.404' | translate }}</h1>-->
-   <h1>ERROR 404</h1>
+   <h1>{{ 'ERRORS.404' | translate }}</h1>
    <div>Visit: {{ config }}</div>
 </div>
