+++ src/index.html
@@ -5,8 +5,8 @@
       <meta charset="UTF-8">
       <meta http-equiv="X-UA-Compatible" content="IE=edge">
       <meta name="viewport" content="width=device-width, initial-scale=1">
-      <!--<link rel="shortcut icon" type="image/png"  href="assets/images/favicon.png">-->
-      <title>test</title>
+      <link rel="shortcut icon" type="image/png"  href="assets/images/favicon.png">
+      <title>Egeo Starter</title>
    </head>
 
    <body>
